+++ src/main/java/org/folio/tm/integration/keycloak/model/ClientAttributes.java
@@ -23,6 +23,8 @@ public class ClientAttributes {
   public static final String CLIENT_SECRET_CREATION_TIME = "client.secret.creation.time";
   public static final String BACKCHANNEL_LOGOUT_SESSION_REQUIRED = "backchannel.logout.session.required";
   public static final String BACKCHANNEL_LOGOUT_REVOKE_OFFLINE_TOKENS = "backchannel.logout.revoke.offline.tokens";
+  public static final String CLIENT_USE_LIGHTWEIGHT_ACCESS_TOKEN_ENABLED =
+    "client.use.lightweight.access.token.enabled";
   public static final String ACCESS_TOKEN_LIFESPAN = "access.token.lifespan";
   public static final String USE_REFRESH_TOKENS = "use.refresh.tokens";
 
@@ -41,6 +43,9 @@ public class ClientAttributes {
   @JsonProperty(BACKCHANNEL_LOGOUT_REVOKE_OFFLINE_TOKENS)
   private boolean backChannelLogoutRevokeOfflineTokens;
 
+  @JsonProperty(CLIENT_USE_LIGHTWEIGHT_ACCESS_TOKEN_ENABLED)
+  private boolean clientUseLightweightAccessTokenEnabled;
+
   @JsonProperty(ACCESS_TOKEN_LIFESPAN)
   private Long accessTokenLifeSpan;
 
@@ -49,12 +54,13 @@ public class ClientAttributes {
 
   public static ClientAttributes defaultValue() {
     var secretTimestamp = Instant.now().getEpochSecond();
-    return new ClientAttributes(false, false, secretTimestamp, true, false, null, null);
+    return new ClientAttributes(false, false, secretTimestamp, true, false, true, null, null);
   }
 
   public static ClientAttributes of(Long accessTokenLifespan, Boolean useRefreshTokens) {
     var secretTimestamp = Instant.now().getEpochSecond();
-    return new ClientAttributes(false, false, secretTimestamp, true, false, accessTokenLifespan, useRefreshTokens);
+    return new ClientAttributes(false, false, secretTimestamp,
+      true, false, true,  accessTokenLifespan, useRefreshTokens);
   }
 
   public Map<String, String> asMap() {
@@ -64,6 +70,7 @@ public class ClientAttributes {
     clientAttributes.put(CLIENT_SECRET_CREATION_TIME, valueOf(clientSecretCreationTime));
     clientAttributes.put(BACKCHANNEL_LOGOUT_SESSION_REQUIRED, valueOf(backChannelLogoutSessionRequired));
     clientAttributes.put(BACKCHANNEL_LOGOUT_REVOKE_OFFLINE_TOKENS, valueOf(backChannelLogoutRevokeOfflineTokens));
+    clientAttributes.put(CLIENT_USE_LIGHTWEIGHT_ACCESS_TOKEN_ENABLED, valueOf(clientUseLightweightAccessTokenEnabled));
 
     applyIfNotNull(accessTokenLifeSpan, value -> clientAttributes.put(ACCESS_TOKEN_LIFESPAN, valueOf(value)));
     applyIfNotNull(useRefreshTokens, value -> clientAttributes.put(USE_REFRESH_TOKENS, valueOf(value)));