+++ src/main/java/org/folio/tm/integration/keycloak/model/ProtocolMapperConfig.java
@@ -19,6 +19,7 @@ public class ProtocolMapperConfig {
   public static final String ID_TOKEN_CLAIM = "id.token.claim";
   public static final String ACCESS_TOKEN_CLAIM = "access.token.claim";
   public static final String USERINFO_TOKEN_CLAIM = "userinfo.token.claim";
+  public static final String LIGHTWEIGHT_CLAIM = "lightweight.claim";
   public static final String USER_ATTRIBUTE = "user.attribute";
   public static final String CLAIM_NAME = "claim.name";
   public static final String JSON_TYPE_LABEL = "jsonType.label";
@@ -32,6 +33,9 @@ public class ProtocolMapperConfig {
   @JsonProperty(USERINFO_TOKEN_CLAIM)
   private boolean userInfoTokenClaim;
 
+  @JsonProperty(LIGHTWEIGHT_CLAIM)
+  private boolean lightweightClaim;
+
   @JsonProperty(USER_ATTRIBUTE)
   private String userAttribute;
 
@@ -41,8 +45,12 @@ public class ProtocolMapperConfig {
   @JsonProperty(JSON_TYPE_LABEL)
   private String jsonTypeLabel;
 
+  public static ProtocolMapperConfig defaultValue() {
+    return new ProtocolMapperConfig(true, true, true, true, null, null, null);
+  }
+
   public static ProtocolMapperConfig forUserAttribute(String userAttribute, String claimName) {
-    return new ProtocolMapperConfig(true, true, true, userAttribute, claimName, STRING_TYPE_LABEL);
+    return new ProtocolMapperConfig(true, true, true, true, userAttribute, claimName, STRING_TYPE_LABEL);
   }
 
   public Map<String, String> asMap() {
@@ -50,6 +58,7 @@ public class ProtocolMapperConfig {
     resultMap.put(ID_TOKEN_CLAIM, valueOf(idTokenClaim));
     resultMap.put(ACCESS_TOKEN_CLAIM, valueOf(accessTokenClaim));
     resultMap.put(USERINFO_TOKEN_CLAIM, valueOf(userInfoTokenClaim));
+    resultMap.put(LIGHTWEIGHT_CLAIM, valueOf(lightweightClaim));
 
     applyIfNotNull(userAttribute, value -> resultMap.put(USER_ATTRIBUTE, value));
     applyIfNotNull(claimName, value -> resultMap.put(CLAIM_NAME, value));