+++ src/main/java/org/folio/tm/integration/keycloak/service/clients/PasswordResetClientService.java
@@ -1,13 +1,13 @@
 package org.folio.tm.integration.keycloak.service.clients;
 
+import static org.folio.common.utils.CollectionUtils.toStream;
 import static org.folio.tm.integration.keycloak.model.Client.OPENID_CONNECT_PROTOCOL;
 import static org.folio.tm.integration.keycloak.service.roles.PasswordResetRoleService.PASSWORD_RESET_ROLE_NAME;
+import static org.folio.tm.integration.keycloak.utils.KeycloakClientUtils.getSubjectProtocolMapper;
 
 import jakarta.persistence.EntityNotFoundException;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.tm.integration.keycloak.KeycloakServerInfoService;
@@ -17,7 +17,6 @@ import org.folio.tm.integration.keycloak.model.ProtocolMapperConfig;
 import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.idm.ProtocolMapperRepresentation;
 import org.keycloak.representations.idm.ProtocolMapperTypeRepresentation;
-import org.keycloak.representations.info.ServerInfoRepresentation;
 
 @Log4j2
 @RequiredArgsConstructor
@@ -65,15 +64,19 @@ public class PasswordResetClientService extends AbstractKeycloakClientService {
 
   @Override
   protected List<ProtocolMapperRepresentation> getProtocolMappers() {
-    return Optional.ofNullable(keycloakServerInfoService.getServerInfo())
-      .map(ServerInfoRepresentation::getProtocolMapperTypes)
-      .map(protocolMapperTypes -> protocolMapperTypes.get(OPENID_CONNECT_PROTOCOL))
-      .stream()
-      .flatMap(Collection::stream)
+    return List.of(getPasswordResetProtocolMapper(), getSubjectProtocolMapper());
+  }
+
+  private ProtocolMapperRepresentation getPasswordResetProtocolMapper() {
+    var serverInfo = keycloakServerInfoService.getServerInfo();
+    if (serverInfo == null || serverInfo.getProtocolMapperTypes() == null) {
+      throw new EntityNotFoundException("Mapper is not found by name: " + PASSWORD_RESET_ACTION_MAPPER);
+    }
+
+    return toStream(serverInfo.getProtocolMapperTypes().get(OPENID_CONNECT_PROTOCOL))
       .filter(mapper -> mapper.getName().equals(PASSWORD_RESET_ACTION_MAPPER))
       .findFirst()
       .map(typeRepr -> toProtocolMapper(typeRepr, null, PASSWORD_RESET_ACTION_ID_CLAIM))
-      .map(List::of)
       .orElseThrow(() -> new EntityNotFoundException("Mapper is not found by name: " + PASSWORD_RESET_ACTION_MAPPER));
   }